requirements.txt
python-telegram-bot==20.0
aiohttp==3.8.4
Pillow==9.4.0
requests==2.28.2
